Rename bridge modules
