+++ tests/integration/test_module1_module3.py
@@ -1,7 +1,7 @@
 """
-Integration Test: Module 1 + Module 3
+Integration Test: Injector + Logging
 -------------------------------------
-Tests the integration between the Injector module and the Logging & Error Handling Layer.
+Tests the integration between the Injector module and the Logging module.
 """
 
 import os
@@ -17,8 +17,8 @@ from typing import Dict, Any
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../src')))
 
 # Import modules to test
-from bridge.modules.module1 import BridgeInjector
-from bridge.modules.module3 import BridgeLogger, ErrorHandler
+from bridge.injector import BridgeInjector
+from bridge.logging import BridgeLogger, ErrorHandler
 
 class TestModule1Module3Integration(unittest.TestCase):
     """Test case for Module 1 and Module 3 integration."""
